animeGAN/requirements.txt
boto3==1.25.5
botocore==1.28.5
coloredlogs==15.0.1
flatbuffers==22.10.26
humanfriendly==10.0
jmespath==1.0.1
mpmath==1.2.1
numpy==1.23.4
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.3.0
protobuf==4.21.9
pyparsing==3.0.9
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
sympy==1.11.1
urllib3==1.26.12